requirements.txt
streamlit==1.50.0
pytrends
undetected-chromedriver
python-dotenv
selenium 
openpyxl
xlwings
setuptools
matplotlib
pyodbc
streamlit-option-menu==0.4.0
streamlit-elements==0.1.0
python-pptx>=0.6.21
psycopg2-binary
psutil